Moving method to GUI
